FEAT: inserção do Periodicos e Eventuais FIX: correções de logs, para melhor a forma de visualizar erros e sucesso alem de separar as logs download, extração e inserção
