+++ requirements.txt
@@ -1,3 +1,6 @@
 beautifulsoup4==4.12.2
 requests==2.31.0
 mysql-connector-python
+bs4
+pandas
+sqlalchemy
